# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sample API Documentation",
        "description": "API documentation for the GreetController",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/gallery": {
            "get": {
                "tags": [
                    "gallery"
                ],
                "summary": "Returns a Sample API gallery response",
                "description": "A sample gallery to test out the API",
                "operationId": "gallery",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil memproses galleries",
                                    "galleries": [
                                        {
                                            "id": 1,
                                            "title": "gallery bell",
                                            "description": "deskripsi gallery bell",
                                            "picture": "bell.jpeg",
                                            "created_at": "2024-11-06T02:20:42.000000Z",
                                            "updated_at": "2024-11-06T02:20:42.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "detail": "strings"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/greet": {
            "get": {
                "tags": [
                    "Greeting"
                ],
                "summary": "Greet the user",
                "description": "Returns a greeting message with user details",
                "operationId": "8b95ba5bd0ea6799e82fed154e0abeba",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "User's first name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "User's last name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil memproses masukan user",
                                    "data": {
                                        "output": "Halo John Doe",
                                        "firstname": "John",
                                        "lastname": "Doe"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "gallery",
            "description": "gallery"
        },
        {
            "name": "Greeting",
            "description": "Greeting"
        }
    ]
}